Overhaul the structure; add lots of material
I've moved most of the sidebar conversation into the "Concepts" structure.
This seems a good way to expand upon a point without dragging out the
narrative. It also serves as a bit of a reference, a mini-guide.

The bulk of the actual coding walkthrough is now taking place in the
"Designing and testing the components" section. In that section there's a
subsection for each of the six application modules. The "Config" class module
is mostly fleshed-out.

This also enables Travis validation of all the code and the site build.