Update Gemfile to maintain Windows compatibility
The update to guides_style_18f v0.3.0 introduce a dependency on
jekyll_pages_api_search. That latter gem depends upon therubyracer, which in
turn depends upon libv8, a gem that manages the binary distribution of the V8
JavaScript library.

jekyll_pages_api_search uses therubyracer to compile the lunr.js search code
and build the index, which is then serialized as `search-index.json`. That gem
also includes glue code to load the index asynchronously from the server.

The problem is, on Windows, there is no existing compiled distribution of the
libv8 gem. For now, we punt on the issue by locking Windows users to
guides_style_18f v0.2.0, the last version that did not depend on
jekyll_pages_api_search.

That said, I've managed to successfully compile libv8 and therubyracer on a
Windows 8 system, after making a few hacks I plan to contribute upstream.
Unfortunately, the resulting library will load and parse JavaScript
expressions just fine, but will cause a segfault when the garbage collector
kicks in (or so that's my current working theory). So I'm tantalizingly close
to getting it to work, but this commit will ensure Windows users can still
build the site and run the `./go` script commands in the meanwhile.